# Models/DTO/PostDTO.cs
namespace BlogsWebApi.Models.DTO
{
    public class PostDTO
    {
        public int PostType { get; set; }
        public string PostPath { get; set; }
        public int IdUser { get; set; }
    }
}