# Models/DTO/UserDTO.cs
namespace BlogsWebApi.Models.DTO
{
    public class UserDTO
    {
        public string Name { get; set; }
        public string LastName { get; set; }
        public string Email { get; set; }
        public string Password { get; set; }
    }
}